# app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PLaylists/Content's API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1"
        }
    ],
    "paths": {
        "/contents": {
            "get": {
                "tags": [
                    "Contents"
                ],
                "summary": "Content List",
                "operationId": "accountValidate",
                "responses": {
                    "200": {
                        "description": "List of Contents"
                    }
                }
            },
            "post": {
                "tags": [
                    "Contents"
                ],
                "summary": "Contents Creation",
                "operationId": "accountValidate",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "url",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "author",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "playlist_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Content created"
                    }
                }
            }
        },
        "/contents/{content}": {
            "put": {
                "tags": [
                    "Contents"
                ],
                "summary": "Content Update",
                "operationId": "content_id",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "url",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "author",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "content",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "playlist_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Content Updated"
                    },
                    "404": {
                        "description": "Content not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Contents"
                ],
                "summary": "Content Delete",
                "operationId": "content_id",
                "parameters": [
                    {
                        "name": "content",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Content Deleted"
                    },
                    "404": {
                        "description": "Content not found"
                    }
                }
            }
        },
        "/playlists": {
            "get": {
                "tags": [
                    "Playlists"
                ],
                "summary": "Playlist List",
                "operationId": "accountValidate",
                "responses": {
                    "200": {
                        "description": "List of Playlists"
                    }
                }
            },
            "post": {
                "tags": [
                    "Playlists"
                ],
                "summary": "Playlist Creation",
                "operationId": "accountValidate",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "author",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Playlist created"
                    }
                }
            }
        },
        "/playlists/{playlist}": {
            "put": {
                "tags": [
                    "Playlists"
                ],
                "summary": "Playlist Update",
                "operationId": "playlist_id",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "author",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "playlist",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Playlist Updated"
                    },
                    "404": {
                        "description": "Playlist not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Playlists"
                ],
                "summary": "Playlist Delete",
                "operationId": "playlist_id",
                "parameters": [
                    {
                        "name": "playlist",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Playlist Deleted"
                    },
                    "404": {
                        "description": "Playlist not found"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}